requirements.txt
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
